requirements.txt
# Database Drivers
psycopg2-binary>=2.9.9
pymongo>=4.6.1

# Data Processing
pandas>=2.2.3
numpy>=1.26.4

# Configuration Management
python-dotenv>=1.0.1

# Optional: For better performance
sqlalchemy>=2.0.25
